Fix: Valgrind errors

+++ gulkov.mikhail/I3/MakeCstring.hpp
@@ -4,6 +4,6 @@
 #include <cstddef>
 #include <istream>
 
-char *make_cstring(char *cstring, unsigned int &size, unsigned int &capacity, std::istream &input);
+char *make_cstring(char *cstring, size_t &size, size_t &capacity, std::istream &input);
 
 #endif
