+++ gulkov.mikhail/I3/ResizeCstring.hpp
@@ -3,6 +3,6 @@
 
 #include <cstddef>
 
-char *check_and_resize(char *arr, unsigned int &size, unsigned int &capacity);
+char *check_and_resize(char *arr, size_t &size, size_t &capacity);
 
 #endif
